Refactor judge aggregation strategy into module
