+++ projects/04-llm-adapter/adapter/core/aggregation.py
@@ -1,32 +1,18 @@
 """応答集約ストラテジ。"""
 from __future__ import annotations
 
-from collections.abc import Mapping, Sequence
-from dataclasses import dataclass, field
-from typing import Any, cast, Protocol, runtime_checkable
+from collections.abc import Sequence
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any, Protocol, cast, runtime_checkable
+
+__path__ = [str(Path(__file__).with_name("aggregation"))]
 
 # 依存は実行時読み込み。型は実体を使う（mypy用に直import）
 try:  # pragma: no cover - 実環境では src.* が存在する
-    from src.llm_adapter.provider_spi import ProviderRequest, ProviderResponse  # type: ignore
+    from src.llm_adapter.provider_spi import ProviderResponse  # type: ignore
 except ModuleNotFoundError:  # pragma: no cover - テスト用フォールバック
     from .providers import ProviderResponse  # type: ignore
-
-    @dataclass(slots=True)
-    class ProviderRequest:  # type: ignore[override]
-        model: str
-        prompt: str = ""
-        messages: Sequence[Mapping[str, Any]] | None = None
-        max_tokens: int | None = None
-        temperature: float | None = None
-        top_p: float | None = None
-        stop: tuple[str, ...] | None = None
-        timeout_s: float | None = None
-        metadata: Mapping[str, Any] | None = None
-        options: dict[str, Any] = field(default_factory=dict)
-
-        @property
-        def prompt_text(self) -> str:
-            return self.prompt
 
 # ===== 基本データ構造 =====
 
@@ -214,110 +200,7 @@ class MaxScoreStrategy:
         )
 
 
-# --- Judge（LLM判定） ---
-
-
-class JudgeStrategy:
-    """
-    LLM ジャッジにより最良を選ぶ。
-    - provider_factory.create(model=...) で判定用プロバイダを作成
-    - 候補を列挙したプロンプトを与え、選択インデックスを抽出
-    """
-
-    name = "judge"
-
-    def __init__(
-        self,
-        *,
-        model: str,
-        provider_factory: Any,
-        prompt_template: str | None = None,
-    ) -> None:
-        self._model = model
-        self._provider_factory = provider_factory
-        self._prompt_template = prompt_template or DEFAULT_JUDGE_TEMPLATE
-
-    def aggregate(
-        self, candidates: Sequence[AggregationCandidate], *, tiebreaker: TieBreaker | None = None
-    ) -> AggregationResult:
-        if not candidates:
-            raise ValueError("judge: candidates must be non-empty")
-
-        # プロンプト生成（1-basedで番号付け）
-        rows: list[str] = []
-        for index, candidate in enumerate(candidates, start=1):
-            raw = candidate.text if candidate.text is not None else candidate.response.text
-            text = raw.strip()
-            rows.append(f"{index}. {text}")
-
-        prompt = self._prompt_template.format(candidates="\n".join(rows))
-
-        # ジャッジプロバイダを作成して問い合わせ
-        judge = self._provider_factory.create(model=self._model)
-        request = ProviderRequest(model=self._model, prompt=prompt, max_tokens=16, temperature=0.0)
-        response: ProviderResponse = judge.invoke(request)
-
-        index_or_none = _parse_choice_index(response.text, total=len(candidates))
-        if index_or_none is None:
-            breaker = tiebreaker or FirstTieBreaker()
-            chosen = breaker.break_tie(candidates)
-            return AggregationResult(
-                chosen=chosen,
-                candidates=list(candidates),
-                strategy=self.name,
-                reason="judge parse failed → tie-break",
-                tie_breaker_used=breaker.name,
-                metadata={"judge_raw": response.text},
-            )
-
-        index = index_or_none
-        chosen = candidates[index]
-        return AggregationResult(
-            chosen=chosen,
-            candidates=list(candidates),
-            strategy=self.name,
-            reason=f"judge selected {index + 1}",
-            tie_breaker_used=None,
-            metadata={"judge_raw": response.text},
-        )
-
-
-DEFAULT_JUDGE_TEMPLATE = (
-    """You are a strict evaluator.
-Read the following candidates and choose the single best answer.
-
-Candidates:
-{candidates}
-
-Rules:
-- Output only the number of the chosen candidate on the first line (e.g., \"2\").
-- Do not add explanations.
-
-Answer with the number only.
-""".strip()
-)
-
-
-def _parse_choice_index(text: str, *, total: int) -> int | None:
-    """
-    返答から 1..total の整数を抽出して 0-based index を返す。
-    先頭行優先、なければ最小の妥当数字を拾う。
-    """
-
-    import re
-
-    if not text:
-        return None
-
-    first_line = text.strip().splitlines()[0]
-    for chunk in (first_line, text):
-        match = re.search(r"\b([1-9][0-9]?)\b", chunk)
-        if not match:
-            continue
-        value = int(match.group(1))
-        if 1 <= value <= total:
-            return value - 1
-    return None
+from .aggregation.judge import DEFAULT_JUDGE_TEMPLATE, JudgeStrategy
 
 
 # 便利ヘルパー：API/CLI から簡単に呼べるように
@@ -333,6 +216,7 @@ __all__ = [
     "MaxScoreTieBreaker",
     "AggregationStrategy",
     "AggregationResolver",
+    "DEFAULT_JUDGE_TEMPLATE",
     "JudgeStrategy",
     "MajorityVoteStrategy",
     "MaxScoreStrategy",
